+++ projects/v2/src/app/modules/tree/tree.functions.ts
@@ -42,9 +42,11 @@ export function makeAS(data: Row[]): Array<AS> {
           id += 1;
 
           if (row.cell_types.length) {
-            newStructure.outdegree.add(
-              `${row.cell_types[0].name}${row.cell_types[0].id}`
-            );
+            row.cell_types.forEach((cell) => {
+              newStructure.outdegree.add(
+                `${cell.name}${cell.id}`
+              );
+            });
           }
           if (i > 0) {
             // needed for the first element to not throw an error
@@ -54,9 +56,11 @@ export function makeAS(data: Row[]): Array<AS> {
           anatomicalStructures.push(newStructure);
         } else {
           if (row.cell_types.length) {
-            anatomicalStructures[foundIndex].outdegree.add(
-              `${row.cell_types[0].name}${row.cell_types[0].id}`
-            );
+            row.cell_types.forEach((cell) => {
+              anatomicalStructures[foundIndex].outdegree.add(
+                `${cell.name}${cell.id}`
+              );
+            });
           }
           if (i > 0) {
             anatomicalStructures[foundIndex].indegree.add(
@@ -193,9 +197,11 @@ export function makeBioMarkers(data: Row[], type?: string): Array<B> {
           };
 
           if (row.cell_types.length) {
-            newStructure.indegree.add(
-              `${row.cell_types[0].name}${row.cell_types[0].id}`
-            );
+            row.cell_types.forEach((cell) => {
+              newStructure.indegree.add(
+                `${cell.name}${cell.id}`
+              );
+            });
           }
 
           bioMarkers.push(newStructure);
@@ -205,9 +211,11 @@ export function makeBioMarkers(data: Row[], type?: string): Array<B> {
             bioMarkers[foundIndex].pathColor = str.color;
           }
           if (row.cell_types.length) {
-            bioMarkers[foundIndex].indegree.add(
-              `${row.cell_types[0].name}${row.cell_types[0].id}`
-            );
+            row.cell_types.forEach((cell) => {
+              bioMarkers[foundIndex].indegree.add(
+                `${cell.name}${cell.id}`
+              );
+            });
           }
         }
       });