+++ projects/v2/src/app/store/sheet.state.ts
@@ -31,6 +31,7 @@ import {
   UpdatePlaygroundData,
   UpdateBottomSheetInfo,
   UpdateBottomSheetDOI,
+  FetchSheetDataFromCSV,
 } from '../actions/sheet.actions';
 import {
   OpenLoading,
@@ -453,16 +454,12 @@ export class SheetState {
     dispatch(new StateReset(TreeState));
     dispatch(new CloseBottomSheet());
     dispatch(new ReportLog(LOG_TYPES.MSG, sheet.display, LOG_ICONS.file));
-
-    const organ: Structure = { name: 'Body', id: '' };
     patchState({ sheet });
     const state = getState();
 
     return this.sheetService.fetchSheetData(sheet.sheetId, sheet.gid).pipe(
       tap((res: ResponseData) => {
-        res.data.forEach((row) => {
-          row.anatomical_structures.unshift(organ);
-        });
+        res.data = this.sheetService.getDataWithBody(res.data);
         setState({
           ...state,
           compareData: [],
@@ -506,6 +503,69 @@ export class SheetState {
     );
   }
 
+  /**
+   * Action to fetch the sheet data for the CSV files. Resets the Sheet State
+   * Accepts the sheet config of the particular sheet
+   */
+  @Action(FetchSheetDataFromCSV)
+  fetchSheetDataFromCSV(
+    { getState, setState, dispatch, patchState }: StateContext<SheetStateModel>,
+    { sheet, url }: FetchSheetDataFromCSV
+  ) {
+    const mode = getState().mode;
+    dispatch(new OpenLoading('Fetching data...'));
+    dispatch(new StateReset(TreeState));
+    dispatch(new CloseBottomSheet());
+    dispatch(new ReportLog(LOG_TYPES.MSG, sheet.display, LOG_ICONS.file));
+
+    patchState({ sheet });
+    const state = getState();
+
+    return this.sheetService.fetchDataFromCSV(url).pipe(
+      tap((res: ResponseData) => {
+        res.data = this.sheetService.getDataWithBody(res.data);
+
+        setState({
+          ...state,
+          compareData: [],
+          compareSheets: [],
+          reportData: [],
+          csv: res.csv,
+          data: res.data,
+          version: 'latest',
+          parsed: res.parsed,
+          mode,
+          sheetConfig: { ...sheet.config, show_ontology: true },
+        });
+
+        dispatch(
+          new ReportLog(
+            LOG_TYPES.MSG,
+            `${sheet.display} data successfully fetched from CSV file url.`,
+            LOG_ICONS.success
+          )
+        );
+
+        dispatch(
+          new UpdateLoadingText(
+            'Fetch data from CSV file successful. Building Visualization..'
+          )
+        );
+      }),
+      catchError((error) => {
+        const err: Error = {
+          msg: `${error.name} (Status: ${error.status})`,
+          hasError: true,
+          status: error.status,
+        };
+        dispatch(
+          new ReportLog(LOG_TYPES.MSG, this.faliureMsg, LOG_ICONS.error)
+        );
+        dispatch(new HasError(err));
+        return of('');
+      })
+    );
+  }
   /**
    * Action to fetch data from assets
    * CURRENTLY DEPRICATED IN V2